Location functions return a move-path and a project-path
This fixes running FileTruck on itself - we need to move to the
move-path, which may be relative, absolute or whatever, we don't
care, but we need to rewrite the project file to take on a relative
path, relative to the project file directory, so it can still find
the files. These might be two different paths so we return them
both separately.